paart/requirements/common.txt
Django==1.4.3
South==0.7.6
Pillow==1.7.8
django-compressor==1.2
cssmin==0.1.4
django-pagination==1.0.7
sh==1.08
Django-Select2==3.3.0
djangorestframework==2.2.1
